+++ packages/zp-data/CHANGELOG.md
@@ -1,5 +1,11 @@
 # zapnode-data
 
+## 0.0.16
+
+### Patch Changes
+
+- change authenticate hook to work without a service
+
 ## 0.0.15
 
 ### Patch Changes